[src/uceil_log2.c] Correction in __gmpfr_ceil_log2, avoiding an incorrect result with tcc: x.s.exp is declared as an unsigned bit-field, so that tcc considers that x.s.exp - 1023 is unsigned. However, since all the values of x.s.exp are representable in an int, according to the integer promotion rules, x.s.exp should be converted to an int, so that the subtraction is signed. So, this appears to be a bug in tcc.
[From SVN r10443 (trunk)]

+++ src/uceil_log2.c
@@ -33,7 +33,7 @@ __gmpfr_ceil_log2 (double d)
   union mpfr_ieee_double_extract x;
 
   x.d = d;
-  exp = x.s.exp - 1023;
+  exp = (long) x.s.exp - 1023;
   x.s.exp = 1023; /* value for 1 <= d < 2 */
   if (x.d != 1.0) /* d: not a power of two? */
     exp++;
